CLEANUP: ebtree: remove a few annoying signedness warnings
(from ebtree 6.0.6)

Care has been taken not to make the code bigger (it even got smaller
due to a possible simplification).
(cherry picked from commit 7a2c1df646049c7daac52677ec11ed63048cd150)

+++ ebtree/ebmbtree.h
@@ -527,7 +527,7 @@ static forceinline struct ebmb_node *__ebmb_lookup_prefix(struct eb_root *root, 
 			 * bits, let's compare prefixes and descend the cover
 			 * subtree if they match.
 			 */
-			if (node->node.bit >> 1 == pfx)
+			if ((unsigned short)node->node.bit >> 1 == pfx)
 				troot = node->node.branches.b[EB_LEFT];
 			else
 				troot = node->node.branches.b[EB_RGHT];
